# frontend/src/lib/api/swagger3.json
{
    "openapi": "3.0.0",
...
    "info": {
        "description": "API системы мониторинга репутации брендов BrandRadar",
        "title": "BrandRadar API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/brands": {
            "get": {
                "description": "Возвращает список всех брендов с пагинацией",
                "tags": [
                    "Brands"
                ],
                "summary": "Список брендов",
                "parameters": [
                    {
                        "description": "Лимит",
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "allOf": [
                                                        {
                                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedData"
                                                        },
                                                        {
                                                            "type": "object",
                                                            "properties": {
                                                                "items": {
                                                                    "type": "array",
                                                                    "items": {
                                                                        "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.BrandResponse"
                                                                    }
                                                                }
                                                            }
                                                        }
                                                    ]
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт новый бренд с ключевыми словами",
                "tags": [
                    "Brands"
                ],
                "summary": "Создать бренд",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.CreateBrandRequest"
                            }
                        }
                    },
                    "description": "Данные бренда",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.BrandResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/brands/{id}": {
            "get": {
                "description": "Возвращает бренд по ID",
                "tags": [
                    "Brands"
                ],
                "summary": "Получить бренд",
                "parameters": [
                    {
                        "description": "UUID бренда",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.BrandResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет данные бренда",
                "tags": [
                    "Brands"
                ],
                "summary": "Обновить бренд",
                "parameters": [
                    {
                        "description": "UUID бренда",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.UpdateBrandRequest"
                            }
                        }
                    },
                    "description": "Обновляемые поля",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.BrandResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет бренд по ID",
                "tags": [
                    "Brands"
                ],
                "summary": "Удалить бренд",
                "parameters": [
                    {
                        "description": "UUID бренда",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1"
        },
        {
            "url": "https://localhost:8080/api/v1"
        }
    ],
    "components": {
        "schemas": {
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.BrandResponse": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "exclusions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "apple pie",
                            "apple fruit"
                        ]
                    },
                    "id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "keywords": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "apple",
                            "iphone",
                            "macbook"
                        ]
                    },
                    "name": {
                        "type": "string",
                        "example": "Apple"
                    },
                    "risk_words": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "scandal",
                            "leak",
                            "bug"
                        ]
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2024-01-01T12:00:00Z"
                    }
                }
            },
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.CreateBrandRequest": {
                "type": "object",
                "required": [
                    "keywords",
                    "name"
                ],
                "properties": {
                    "exclusions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "apple pie",
                            "apple fruit"
                        ]
                    },
                    "keywords": {
                        "type": "array",
                        "minItems": 1,
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "apple",
                            "iphone",
                            "macbook"
                        ]
                    },
                    "name": {
                        "type": "string",
                        "example": "Apple"
                    },
                    "risk_words": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "scandal",
                            "leak",
                            "bug"
                        ]
                    }
                }
            },
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.ErrorDetail": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "NOT_FOUND"
                    },
                    "message": {
                        "type": "string",
                        "example": "resource not found"
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse": {
                "type": "object",
                "properties": {
                    "data": {},
                    "error": {
                        "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorDetail"
                    }
                }
            },
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedData": {
                "type": "object",
                "properties": {
                    "items": {},
                    "limit": {
                        "type": "integer",
                        "example": 20
                    },
                    "offset": {
                        "type": "integer",
                        "example": 0
                    },
                    "total": {
                        "type": "integer",
                        "example": 100
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedData"
                    },
                    "error": {
                        "type": "string",
                        "example": "null"
                    }
                }
            },
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse": {
                "type": "object",
                "properties": {
                    "data": {},
                    "error": {
                        "type": "string",
                        "example": "null"
                    }
                }
            },
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.UpdateBrandRequest": {
                "type": "object",
                "properties": {
                    "exclusions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "apple pie"
                        ]
                    },
                    "keywords": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "apple",
                            "iphone"
                        ]
                    },
                    "name": {
                        "type": "string",
                        "example": "Apple Inc"
                    },
                    "risk_words": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "scandal"
                        ]
                    }
                }
            }
        }
    }
}